fix BufLeave (broke exiting from CtrlP)
Although "normal! i" in BufLeave should run in terminal buffers only it somehow breaks CtrlP and maybe other plugins in a way that its window could not easily be closed again and some other mappings didn't work either afterwards.

Replacing "normal! i" with "startinsert" or "feedkeys" doesn't work and operates on the next buffer entered instead of the one about to be left. Wrapping it in "execute" fixed this.